+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sim/spawners/from_files/from_files.py
@@ -57,43 +57,8 @@ def spawn_from_usd(
     Raises:
         FileNotFoundError: If the USD file does not exist at the given path.
     """
-    # check file path exists
-    if not check_file_path(cfg.usd_path):
-        raise FileNotFoundError(f"USD file not found at path: '{cfg.usd_path}'.")
-    # spawn asset if it doesn't exist.
-    if not prim_utils.is_prim_path_valid(prim_path):
-        # add prim as reference to stage
-        prim_utils.create_prim(
-            prim_path,
-            usd_path=cfg.usd_path,
-            translation=translation,
-            orientation=orientation,
-            scale=cfg.scale,
-        )
-    else:
-        carb.log_warn(f"A prim already exists at prim path: '{prim_path}'.")
-
-    # modify rigid body properties
-    if cfg.rigid_props is not None:
-        schemas.modify_rigid_body_properties(prim_path, cfg.rigid_props)
-    # modify collision properties
-    if cfg.collision_props is not None:
-        schemas.modify_collision_properties(prim_path, cfg.collision_props)
-    # modify articulation root properties
-    if cfg.articulation_props is not None:
-        schemas.modify_articulation_root_properties(prim_path, cfg.articulation_props)
-    # apply visual material
-    if cfg.visual_material is not None:
-        if not cfg.visual_material_path.startswith("/"):
-            material_path = f"{prim_path}/{cfg.visual_material_path}"
-        else:
-            material_path = cfg.visual_material_path
-        # create material
-        cfg.visual_material.func(material_path, cfg.visual_material)
-        # apply material
-        bind_visual_material(prim_path, material_path)
-    # return the prim
-    return prim_utils.get_prim_at_path(prim_path)
+    # spawn asset from the given usd file
+    return _spawn_from_usd_file(prim_path, cfg.usd_path, cfg, translation, orientation)
 
 
 @clone
@@ -132,42 +97,10 @@ def spawn_from_urdf(
     Raises:
         FileNotFoundError: If the URDF file does not exist at the given path.
     """
-    # spawn asset if it doesn't exist.
-    if not prim_utils.is_prim_path_valid(prim_path):
-        # urdf loader
-        urdf_loader = converters.UrdfConverter(cfg)
-        # add prim as reference to stage
-        prim_utils.create_prim(
-            prim_path,
-            usd_path=urdf_loader.usd_path,
-            translation=translation,
-            orientation=orientation,
-            scale=cfg.scale,
-        )
-    else:
-        carb.log_warn(f"A prim already exists at prim path: '{prim_path}'. Skipping...")
-
-    # modify rigid body properties
-    if cfg.rigid_props is not None:
-        schemas.modify_rigid_body_properties(prim_path, cfg.rigid_props)
-    # modify collision properties
-    if cfg.collision_props is not None:
-        schemas.modify_collision_properties(prim_path, cfg.collision_props)
-    # modify articulation root properties
-    if cfg.articulation_props is not None:
-        schemas.modify_articulation_root_properties(prim_path, cfg.articulation_props)
-    # apply visual material
-    if cfg.visual_material is not None:
-        if not cfg.visual_material_path.startswith("/"):
-            material_path = f"{prim_path}/{cfg.visual_material_path}"
-        else:
-            material_path = cfg.visual_material_path
-        # create material
-        cfg.visual_material.func(material_path, cfg.visual_material)
-        # apply material
-        bind_visual_material(prim_path, material_path)
-    # return the prim
-    return prim_utils.get_prim_at_path(prim_path)
+    # urdf loader to convert urdf to usd
+    urdf_loader = converters.UrdfConverter(cfg)
+    # spawn asset from the generated usd file
+    return _spawn_from_usd_file(prim_path, urdf_loader.usd_path, cfg, translation, orientation)
 
 
 def spawn_ground_plane(
@@ -251,6 +184,83 @@ def spawn_ground_plane(
     # Remove the light from the ground plane
     # It isn't bright enough and messes up with the user's lighting settings
     omni.kit.commands.execute("ToggleVisibilitySelectedPrims", selected_paths=[f"{prim_path}/SphereLight"])
+
+    # return the prim
+    return prim_utils.get_prim_at_path(prim_path)
+
+
+"""
+Helper functions.
+"""
+
+
+def _spawn_from_usd_file(
+    prim_path: str,
+    usd_path: str,
+    cfg: from_files_cfg.FileCfg,
+    translation: tuple[float, float, float] | None = None,
+    orientation: tuple[float, float, float, float] | None = None,
+) -> Usd.Prim:
+    """Spawn an asset from a USD file and override the settings with the given config.
+
+    In case a prim already exists at the given prim path, then the function does not create a new prim
+    or throw an error that the prim already exists. Instead, it just takes the existing prim and overrides
+    the settings with the given config.
+
+    Args:
+        prim_path: The prim path or pattern to spawn the asset at. If the prim path is a regex pattern,
+            then the asset is spawned at all the matching prim paths.
+        cfg: The configuration instance.
+        translation: The translation to apply to the prim w.r.t. its parent prim. Defaults to None, in which
+            case the translation specified in the generated USD file is used.
+        orientation: The orientation in (w, x, y, z) to apply to the prim w.r.t. its parent prim. Defaults to None,
+            in which case the orientation specified in the generated USD file is used.
+
+    Returns:
+        The prim of the spawned asset.
+
+    Raises:
+        FileNotFoundError: If the USD file does not exist at the given path.
+    """
+    # check file path exists
+    if not check_file_path(usd_path):
+        raise FileNotFoundError(f"USD file not found at path: '{usd_path}'.")
+    # spawn asset if it doesn't exist.
+    if not prim_utils.is_prim_path_valid(prim_path):
+        # add prim as reference to stage
+        prim_utils.create_prim(
+            prim_path,
+            usd_path=usd_path,
+            translation=translation,
+            orientation=orientation,
+            scale=cfg.scale,
+        )
+    else:
+        carb.log_warn(f"A prim already exists at prim path: '{prim_path}'.")
+
+    # modify rigid body properties
+    if cfg.rigid_props is not None:
+        schemas.modify_rigid_body_properties(prim_path, cfg.rigid_props)
+    # modify collision properties
+    if cfg.collision_props is not None:
+        schemas.modify_collision_properties(prim_path, cfg.collision_props)
+    # modify mass properties
+    if cfg.mass_props is not None:
+        schemas.modify_mass_properties(prim_path, cfg.mass_props)
+    # modify articulation root properties
+    if cfg.articulation_props is not None:
+        schemas.modify_articulation_root_properties(prim_path, cfg.articulation_props)
+
+    # apply visual material
+    if cfg.visual_material is not None:
+        if not cfg.visual_material_path.startswith("/"):
+            material_path = f"{prim_path}/{cfg.visual_material_path}"
+        else:
+            material_path = cfg.visual_material_path
+        # create material
+        cfg.visual_material.func(material_path, cfg.visual_material)
+        # apply material
+        bind_visual_material(prim_path, material_path)
 
     # return the prim
     return prim_utils.get_prim_at_path(prim_path)